Refactor code + files
