+++ src/components/ManageUsers/EditUserModal/EditUserModal.module.css
@@ -1,0 +1,32 @@
+.container {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+}
+
+.container h1 {
+  font-weight: bold;
+  margin-bottom: 10px;
+}
+
+.container form {
+  width: 280px;
+}
+
+.container span {
+  text-align: center;
+}
+
+.container label {
+  font-weight: 500;
+  font-size: 20px !important;
+  line-height: 22px;
+}
+
+.container button {
+  border-radius: 3px;
+  height: 40px;
+  font-weight: 500;
+  transform: translateY(-25px);
+}
