+++ src/components/ManageUsers/ManageUsers.module.css
@@ -1,40 +1,16 @@
-.container {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-}
-
 .header-title {
   display: flex;
-  flex-direction:row;
+  flex-direction: row;
   justify-content: center;
   align-items: center;
   column-gap: 2em;
 }
 
-.container h1 {
-  font-weight: bold;
-  margin-bottom: 10px;
+.container {
+  display: flex;
+  justify-content: center;
 }
 
-.container form {
-  width: 280px;
-}
-
-.container span {
-  text-align: center;
-}
-
-.container label {
-  font-weight: 500;
-  font-size: 20px !important;
-  line-height: 22px;
-}
-
-.container button {
-  border-radius: 3px;
-  height: 40px;
-  font-weight: 500;
-  transform: translateY(-25px);
+.page {
+  width: 75%;
 }
