+++ src/pages/ManageUsers/ManageUsersPage.module.css
@@ -3,8 +3,6 @@
   justify-content: center;
 }
 
-.page{
-  width:75%;
+.page {
+  width: 75%;
 }
-
-
